Add osfs.ToOSPath() to convert FS paths to OS paths
